알라딘 판매책 필드 수
