+++ src/main/java/com/nhnacademy/book/book/controller/AladinApiController.java
@@ -1,6 +1,5 @@
 package com.nhnacademy.book.book.controller;
 
-//import com.nhnacademy.book.book.service.AladinApiService;
 import com.nhnacademy.book.book.service.api.ApiService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -46,16 +45,4 @@ public class AladinApiController {
     }
 
 
-    /**
-     * 특정 ISBN(ItemId)를 기준으로 데이터베이스에 저장
-     */
-    @PostMapping("/sync/itemid/{isbn13}")
-    public ResponseEntity<String> syncBookByIsbn(@PathVariable String isbn13) {
-        boolean isSaved = aladinApiService.saveBooksByIsbns(List.of(isbn13));
-        if (isSaved) {
-            return ResponseEntity.ok("Book with ISBN " + isbn13 + " has been successfully saved.");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No valid data found for ISBN " + isbn13 + ".");
-        }
-    }
 }
